# Conflicts:
#	crates/core/desis/src/runtime.rs
#	crates/core/desis/src/schema.rs
#	crates/core/gemfactory/src/tests.rs
#	crates/core/intex/src/certified.rs
#	crates/core/intex/src/schema.rs
#	crates/core/intex/src/tests.rs
#	crates/core/intexfactory/src/runtime.rs
#	crates/core/intexfactory/src/schema.rs
#	crates/core/intexfactory/src/tests.rs
#	crates/core/metadosis/src/tests/ocomp_request.rs

+++ crates/core/metadosis/src/settlement.rs
@@ -263,7 +263,7 @@ fn dispatch_brief(
     let brief_supply = if is_green { supply } else { U256::ZERO };
     let receipt = outbe_desis::api::dispatch_auction_brief(
         ctx.storage.clone(),
-        u32::from(wwd),
+        wwd,
         brief_supply,
         reference_prices,
         is_green,
